requirements.txt
flask>=3.0.0
jinja2>=3.1.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
requests>=2.28.0
sqlalchemy
flask-sqlalchemy
python-dotenv
pyrebase4
